requirements.txt
# /home/smartgrow/requirements.txt
# Встановлення: pip install -r requirements.txt

# ── Raspberry Pi / GPIO ──────────────────────────────────────
RPi.GPIO>=0.7.1
adafruit-blinka>=8.0.0
adafruit-circuitpython-ads1x15>=2.3.0
adafruit-circuitpython-dht>=4.0.0
smbus2>=0.4.0
spidev>=3.6

# ── Дисплеї ──────────────────────────────────────────────────
luma.oled>=3.13.0
luma.lcd>=2.9.0
Pillow>=10.0.0

# ── Веб-сервер ───────────────────────────────────────────────
flask>=3.0.0
flask-socketio>=5.3.0
gevent>=23.0.0
gevent-websocket>=0.10.1

# ── Telegram ─────────────────────────────────────────────────
python-telegram-bot==20.7

# ── LangChain AI агент ───────────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.30.0

# ── Утиліти ──────────────────────────────────────────────────
requests>=2.31.0
